Update the namespace of all the classes
