+++ src/Diduhless/Parties/event/PartiesEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Diduhless\Parties\event;
+
+
+use pocketmine\event\Event;
+
+abstract class PartiesEvent extends Event {
+
+}
