+++ src/Diduhless/Parties/event/party/PartyCreateEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Diduhless\Parties\event\party;
+
+
+class PartyCreateEvent extends PartyEvent {
+
+}
