+++ src/Diduhless/Parties/event/party/PartyDisbandEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Diduhless\Parties\event\party;
+
+
+class PartyDisbandEvent extends PartyEvent {
+
+}
