+++ src/Diduhless/Parties/event/party/PartyPromoteEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Diduhless\Parties\event\party;
+
+
+class PartyPromoteEvent extends PartyEvent {
+
+}
