+++ src/Diduhless/Parties/event/session/SessionCloseEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Diduhless\Parties\event\session;
+
+
+class SessionCloseEvent extends SessionEvent {
+
+}
