+++ src/Diduhless/Parties/event/session/SessionOpenEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Diduhless\Parties\event\session;
+
+
+class SessionOpenEvent extends SessionEvent {
+
+}
