--- src/Parties/event/PartiesEvent.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Parties\event;
-
-
-use pocketmine\event\Event;
-
-abstract class PartiesEvent extends Event {
-
-}
