--- src/Parties/event/party/PartyCreateEvent.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Parties\event\party;
-
-
-class PartyCreateEvent extends PartyEvent {
-
-}
