--- src/Parties/event/party/PartyPromoteEvent.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Parties\event\party;
-
-
-class PartyPromoteEvent extends PartyEvent {
-
-}
