--- src/Parties/event/session/SessionCloseEvent.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Parties\event\session;
-
-
-class SessionCloseEvent extends SessionEvent {
-
-}
