--- src/Parties/event/session/SessionOpenEvent.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Parties\event\session;
-
-
-class SessionOpenEvent extends SessionEvent {
-
-}
